# back/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TECLAR — API",
    "version": "1.0",
    "description": "Gerada do mesmo contrato que o CONTRATO-API.md (js/nucleo/api.contrato.ts), por gerar-contrato.mjs. Rotas marcadas \"(ainda não implementada)\" ainda não existem no back.\n\nPara testar uma rota com cadeado: faça o POST /api/auth/login, copie o \"token\" da resposta e cole em **Authorize** (sem a palavra Bearer)."
  },
...
  "paths": {
...
    "/api/exercicios": {
      "get": {
        "tags": [
          "Biblioteca de exercícios (professor)"
        ],
        "operationId": "exercicios.listar",
        "summary": "listar: Paginado<Exercicio> — paginado no contrato; quem lê passa por desembrulhar()",
        "description": "**Chamada no front:** `api.exercicios.listar`\n\n**Resposta:** 200 Paginado<Exercicio> — paginado no contrato; quem lê passa por desembrulhar()\n\n**Identidade:**\n\nO token: WHERE ProfessorID = conta do token.\n\n**Regras de negócio no back (a tela não calcula):**\n\nSó os ativos (Ativo = TRUE). atribuidoA = em quantas turmas está atribuído (COUNT em AtribuicoesProf). A tela não conta.\n\n**Notas:**\n\nSem tamanho de página definido: vem tudo, com pagina = 1.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Paginado<Exercicio> — paginado no contrato; quem lê passa por desembrulhar()",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "Exemplo 1",
                    "value": {
                      "total": 4,
                      "pagina": 1,
                      "itens": [
                        {
                          "id": "ex-prof-1",
                          "professorId": "u-2",
                          "titulo": "Acentuação em foco",
                          "texto": "...",
                          "dificuldade": "medio",
                          "tempoLimiteSegundos": 120,
                          "atribuidoA": 2
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "`TIPO_INVALIDO` — token de aluno",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TIPO_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TIPO_INVALIDO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Biblioteca de exercícios (professor)"
        ],
        "operationId": "exercicios.criar",
        "summary": "criar: Exercicio",
        "description": "**Chamada no front:** `api.exercicios.criar`\n\n**Corpo:** DadosExercicio — o formulário inteiro\n\n**Resposta:** 200 Exercicio\n\n**Identidade:**\n\nO token vira o ProfessorID. professorId no corpo é recusado com 400.\n\n**Regras de negócio no back (a tela não calcula):**\n\nOs limites são os do front (LIMITES em js/utils/validacao.ts). Quebra de linha no texto vira espaço antes de contar e gravar, como a tela faz.\n\n**Notas:**\n\ntempoLimiteSegundos 0 = sem limite. A contagem de caracteres não tem coluna: a tela conta do texto.",
        "parameters": [],
        "requestBody": {
          "description": "DadosExercicio — o formulário inteiro",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "examples": {
                "exemplo1": {
                  "summary": "Exemplo 1",
                  "value": {
                    "titulo": "Pontuação e ritmo",
                    "texto": "Vírgula, ponto; dois-pontos: ...",
                    "dificuldade": "facil",
                    "tempoLimiteSegundos": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercicio",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "Exemplo 1",
                    "value": {
                      "id": "ex-prof-8",
                      "professorId": "u-2",
                      "titulo": "Pontuação e ritmo",
                      "texto": "Vírgula, ponto; dois-pontos: ...",
                      "dificuldade": "facil",
                      "tempoLimiteSegundos": 0,
                      "atribuidoA": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "`DADOS_INVALIDOS` — título fora de 3 a 100 caracteres, texto fora de 20 a 2000, dificuldade fora de facil/medio/dificil, tempoLimiteSegundos que não é inteiro de 0 a 3600, ou o corpo traz professorId",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "DADOS_INVALIDOS",
                    "value": {
                      "mensagem": "...",
                      "codigo": "DADOS_INVALIDOS"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "`TIPO_INVALIDO` — token de aluno",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TIPO_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TIPO_INVALIDO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/exercicios/{id}": {
      "get": {
        "tags": [
          "Biblioteca de exercícios (professor)"
        ],
        "operationId": "exercicios.obter",
        "summary": "obter: ExercicioDetalhe — mesma forma de um item da listagem",
        "description": "**Chamada no front:** `api.exercicios.obter`\n\n**Resposta:** 200 ExercicioDetalhe — mesma forma de um item da listagem\n\n**Identidade:**\n\nConta: o exercício tem de ser dela; ?turma= é ignorado. Aluno: ?turma= é a turma em que ele está treinando, que tem de ser a dele (Alunos.ClassID), e o exercício tem de estar atribuído a ela.\n\n**Notas:**\n\nA mesma regra vale no POST /sessoes, para o aluno descobrir na ABERTURA do treino, e não depois de digitar o texto inteiro.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "turma",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ExercicioDetalhe — mesma forma de um item da listagem",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "Exemplo 1",
                    "value": {
                      "id": "ex-prof-1",
                      "professorId": "u-2",
                      "titulo": "Acentuação em foco",
                      "texto": "...",
                      "dificuldade": "medio",
                      "tempoLimiteSegundos": 120,
                      "atribuidoA": 2
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — não existe, é de outra conta, está arquivado, ou (aluno) não está atribuído à turma dele",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Biblioteca de exercícios (professor)"
        ],
        "operationId": "exercicios.atualizar",
        "summary": "atualizar: Exercicio — como ficou",
        "description": "**Chamada no front:** `api.exercicios.atualizar`\n\n**Corpo:** DadosExercicio — o formulário inteiro\n\n**Resposta:** 200 Exercicio — como ficou\n\n**Identidade:**\n\nO token: o exercício tem de ser da conta. professorId no corpo é recusado com 400.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "DadosExercicio — o formulário inteiro",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "examples": {
                "exemplo1": {
                  "summary": "Exemplo 1",
                  "value": {
                    "titulo": "Pontuação e ritmo",
                    "texto": "...",
                    "dificuldade": "medio",
                    "tempoLimiteSegundos": 90
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercicio — como ficou",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "Exemplo 1",
                    "value": {
                      "id": "ex-prof-2",
                      "professorId": "u-2",
                      "titulo": "Pontuação e ritmo",
                      "texto": "...",
                      "dificuldade": "medio",
                      "tempoLimiteSegundos": 90,
                      "atribuidoA": 2
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "`DADOS_INVALIDOS` — as mesmas regras do POST /exercicios; os quatro campos são obrigatórios",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "DADOS_INVALIDOS",
                    "value": {
                      "mensagem": "...",
                      "codigo": "DADOS_INVALIDOS"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — não existe, é de outra conta ou está arquivado",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Biblioteca de exercícios (professor)"
        ],
        "operationId": "exercicios.excluir",
        "summary": "excluir: (sem corpo)",
        "description": "**Chamada no front:** `api.exercicios.excluir`\n\n**Resposta:** 204 (sem corpo)\n\n**Identidade:**\n\nO token: o exercício tem de ser da conta.\n\n**Regras de negócio no back (a tela não calcula):**\n\nARQUIVA, não apaga: grava ExerciciosProf.Ativo = false e apaga as linhas de AtribuicoesProf dele, numa transação. Ele some da biblioteca e de todas as turmas; a linha de ExerciciosProf e as sessões já feitas continuam, então histórico e relatórios não mudam.\n\n**Notas:**\n\nNão é um DELETE de verdade porque o ON DELETE CASCADE de SessionsProf levaria as sessões dos alunos, que são o trabalho deles, não do professor. Não há rota para desarquivar. A tela avisa antes de confirmar.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "(sem corpo)"
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — não existe, é de outra conta ou já está arquivado",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}